feat: Add nested accordion menu
